Fixing the calculations for week domain lists
With domain="year" and subdomain="week", years like 2012 won't correctly
fetch the list of weeks. Since 12/31 is the start of a week, and `stop`
was set to 12/31 as well, d3.time.mondays/sundays would not give 12/31 in
the list since it is not inclusive.

Therefore, I've added a single microsecond to `stop`, this allowed it to
retrieve the correct values for 2012.